pyproject.toml
[project]
name = "langchain-course-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.16",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]
